requirements.txt
Telethon==1.26.1
cryptg==0.4.0
pyrogram
TgCrypto
pytz==2022.7
dnspython==2.2.1
motor==3.1.1
pymongo==4.3.3
python-dotenv==0.21.0
psutil==5.9.4
aiohttp==3.8.3
aiofiles==22.1.0
speedtest-cli==2.1.3
aria2p==0.11.3
requests==2.28.2
heroku3==5.2.0
python-magic==0.4.27
aiofiles
flask
gunicorn